+++ illico/ovo/__init__.py
@@ -1,10 +1,14 @@
-from illico.ovo.dense_ovo import dense_ovo_mwu_kernel_over_contiguous_col_chunk
+from illico.ovo.dense_ovo import (
+    dense_ovo_mwu_kernel,
+    dense_ovo_mwu_kernel_over_contiguous_col_chunk,
+)
 from illico.ovo.sparse_ovo import (
     csc_ovo_mwu_kernel_over_contiguous_col_chunk,
     csr_ovo_mwu_kernel_over_contiguous_col_chunk,
 )
 
 __all__ = [
+    "dense_ovo_mwu_kernel",
     "dense_ovo_mwu_kernel_over_contiguous_col_chunk",
     "csc_ovo_mwu_kernel_over_contiguous_col_chunk",
     "csr_ovo_mwu_kernel_over_contiguous_col_chunk",